This is a miscellaneous set of fixes necessary to build on WINNT.
Some of thefixes are errors in the non-WINNT specific areas related to
type mismatches andcasting issues.

+++ libntp/dolfptoa.c
@@ -13,7 +13,7 @@ dolfptoa(
 	u_long fpi,
 	u_long fpv,
 	int neg,
-	int ndec,
+	short ndec,
 	int msec
 	)
 {
